# Add `<list-of>` self-self cast optimization